new file structure without submodules. Inventory service wip
