+++ main.py
@@ -1,14 +1,13 @@
+import uvicorn
 from sqlmodel import create_engine, Session, SQLModel
 from contextlib import asynccontextmanager
 from fastapi import FastAPI, Depends
 from typing import Annotated
 from datetime import date
-from .inventory.repositories.books_repository import BooksRepository
-from .inventory.inventory_service import InventoryService
-from .inventory.models.book import Book 
-from .inventory.models.author import Author
-from .inventory.models.genre import Genre
-from .inventory.models.publisher import Publisher
+from inventory.repositories import BooksRepository
+from inventory.inventory_service import InventoryService
+from inventory.models import Book, BookResponse, Author, Genre, Publisher
+
 
 
 sql_url = "postgresql://postgres:admin@localhost:5431/db"
@@ -40,14 +39,15 @@ rep = BooksRepository(engine)
 service = InventoryService()
 
 
-@app.get("/books", response_model=list[Book])
+@app.get("/books", response_model=list[BookResponse])
 async def get_books():
-    return rep.get_all()
+    books = rep.get_all()
+    #print (books)
+    return books
 
 @app.get("/book/{book_isbn}", response_model=Book)
 async def get_book_by_isbn(book_isbn: int):
-    rep.get_by_isbn(book_isbn)
-    return {"book_isbn": book_isbn}
+    return rep.get_by_isbn(book_isbn)
 
 @app.delete("/book", response_model=Book)
 async def delete_book_by_isbn(book_isbn: int):
@@ -65,3 +65,6 @@ async def update_book(book: Book):
     service.new_genre(book)
     rep.update(book)
     return book
+
+if __name__=="__main__":
+    uvicorn.run(app, host="0.0.0.0", port=8000)
